Subscription model change [Drew]
Subscriptions now a part of the data_store as opposed to existing in
both coffee and customer. This change was made up further reflection/
interpretation of spec tests. Under this model, there is only one place
to store subscription information. This mirrors a situation where
subscribers are stored in a database as opposed to being the
responsibility of coffee or customer objects